feat: Cliente repo y controlador

+++ app/src/cliente/infrastructure/mapper/cliente_orm_mapper.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Integer, CheckConstraint, ForeignKey
+from sqlalchemy import Column, String, BigInteger, CheckConstraint, ForeignKey
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 from app.src.common.infrastructure.sqlalchemy_base import Base
@@ -11,9 +11,9 @@ class ClienteOrm(Base):
     # Columnas:
     id_cliente = Column(UUID(as_uuid=True), primary_key=True)
     nombre_cliente = Column(String(100), nullable=False)
-    numero_cedula = Column(Integer, CheckConstraint('numero_cedula >= 0'), nullable=False, unique=True)
+    numero_cedula = Column(BigInteger, CheckConstraint('numero_cedula >= 0'), nullable=False, unique=True)
     email_cliente = Column(String(50), nullable=False, unique=True)
-    numero_telefono = Column(Integer, CheckConstraint('numero_telefono >= 0'))
+    numero_telefono = Column(BigInteger, CheckConstraint('numero_telefono >= 0'))
     user_id = Column(UUID(as_uuid=True), ForeignKey('user.user_id'), nullable=False, unique=True)
     
     # Relaciones:
